The yellow circle properly disappears when releasing 'V'
Prior state: when releasing the 'V' key, the yellow circle would only disappear after the mouse moved.
New state: when releasing the 'V' key, the yellow circle immediately disappears.

Also factored out the name of the key used to activate the plugin, to make it more obvious how to change it if needed.